Removed and fixed many pytype ignores.
This also involved removing code that was required for Python 3.7 compatibility, as some type errors came from that code.

Added the new `Self` from typing/typing_extensions to simplify types.

Removed forward declaration of `Optimizer.State`.

PiperOrigin-RevId: 623821712

+++ kfac_jax/_src/utils/misc.py
@@ -298,7 +298,7 @@ def auto_scope_function(func: Callable[..., T]) -> Callable[..., T]:
   return wrapped
 
 
-def default_batch_size_extractor(batch: types.Batch) -> Numeric:
+def default_batch_size_extractor(batch: types.Batch) -> int:
   """Computes the batch size as the size of axis `0` of the first element."""
   return jax.tree_util.tree_leaves(batch)[0].shape[0]
